+++ src/core/metadata/controller.ts
@@ -14,24 +14,10 @@
  */
 
 import 'reflect-metadata';
-import * as http from 'http';
+import { hasEntryPoint } from './entry';
 
-/**
- * Kite controller metadata
- */
 export interface ControllerMetadata {
-    /**
-     * which method to receive data
-     * 
-     * @type { string }
-     */
-    method?: string;
-
-    /**
-     * privilege of accessing this controller
-     * @type { string | number }
-     */
-    privilege?: string | number;
+    [name: string]: any
 }
 
 /**
@@ -58,33 +44,14 @@ export interface ControllerMetadata {
  * 
  * A Kite controller class must have one (only one) entry point, which must be annotated with an
  * `@Entry` decorator, Kite will locate this entry point function and invoke it when request comes.
- * 
- * ### Metadata properties
- * + __method__ - which HTTP method ("GET", "POST" ...) to receive client data. If this value is set
- *  client request is restricted to this method, any other methods will cause an error. For example, 
- *  `@Controller({method: "GET"})` tells Kite this controller only accepts "GET" method on request, 
- *  request with "POST" will get an error.
- * + __privilege__ - access privilege of this controller. #see Kite privilege mechanism#
- * 
- * If none of these properties is provided, the controller will:
- * + accept any valid HTTP method from client request
- * + any one can invoke, no access check
  */
-export function Controller(metadata?: ControllerMetadata) {
+export function Controller<T extends ControllerMetadata>(metadata?: T) {
     return function (constructor: Function) {
-        if (!Reflect.hasMetadata('kite:entrypoint', constructor.prototype)) {
+        if (!hasEntryPoint(constructor.prototype)) {
             throw new Error(`Missing entry point for controller ${constructor.name}`);
         }
 
-        metadata = metadata || {};
-        if (metadata.method) {
-            let accept = metadata.method.toUpperCase();
-            if (!http.METHODS.includes(accept)) {
-                throw new Error(`Invalid method "${accept}", support methods: ${http.METHODS.toLocaleString()}`);
-            }
-        }
-
-        Reflect.defineMetadata('kite:controller', metadata, constructor);
+        Reflect.defineMetadata('kite:controller', metadata || {}, constructor);
     }
 }