+++ src/core/metadata/index.ts
@@ -1,4 +1,4 @@
-export { Controller, ControllerMetadata } from './controller';
+export { Controller, ControllerMetadata, getControllerMetadata } from './controller';
 export * from './entry';
 export { Inject } from './inject';
 export { Injectable } from './injectable';