+++ src/core/types/config.ts
@@ -14,7 +14,6 @@
  */
 
 import { Router, RouterProvider } from './router';
-import { HolderClass } from './holder';
 import * as http from 'http';
 import { ParserProvider } from './parser';
 import { Responder } from './responder';
@@ -118,7 +117,7 @@ export interface Config {
     /**
      * Holder class
      */
-    holder?: HolderClass;
+    holderClass?: { new(...args: any[]): any };
 
     /**
      * Router provider