+++ src/core/types/index.ts
@@ -14,9 +14,7 @@
  */
 
 export * from './config';
-export * from './context';
 export * from './init';
-export * from './holder';
 export * from './router';
 export * from './parser';
 export * from './responder';